Refactor return value of loadPlanetsData function
The changes refactor the return of the loadPlanetsData function to provide more detailed information. The old return value just listed the planets, now the function returns an array of objects that includes the planet's name, the number of moons and its star's name for each planet. Also, unused commented codes are removed.